+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - The error-path syntax is highlighted
   - When the cursor is on a line that begins with an error-path syntax, pressing `<CR>` will go to that error's file, row and column
   - If the error's file is not found within the current path, the user is prompted to enter the directory to search within
+- The ability to use `:aboveleft`, `:belowright`, `:topleft` and `:botright` with the `:Compile` and `:Recompile` commands
+  - Also the ability to pass `smods.split` to the API functions
 
 ## [1.0.2] - 2023-10-29
 
@@ -23,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - The ability to use `:silent` with the `:Compile` and `:Recompile` commands
-  - Also the ability to pass the `smods.silent` to the API functions
+  - Also the ability to pass `smods.silent` to the API functions
 
 ## [1.0.1] - 2023-10-27